+++ app/src/main/java/eu/kanade/tachiyomi/data/database/queries/MangaQueries.kt
@@ -9,10 +9,6 @@ import eu.kanade.tachiyomi.data.database.models.Manga
 import eu.kanade.tachiyomi.data.database.models.SourceIdMangaCount
 import eu.kanade.tachiyomi.data.database.resolvers.MangaDateAddedPutResolver
 import eu.kanade.tachiyomi.data.database.resolvers.MangaFavoritePutResolver
-import eu.kanade.tachiyomi.data.database.resolvers.MangaFilteredScanlatorsPutResolver
-import eu.kanade.tachiyomi.data.database.resolvers.MangaFlagsPutResolver
-import eu.kanade.tachiyomi.data.database.resolvers.MangaInfoPutResolver
-import eu.kanade.tachiyomi.data.database.resolvers.MangaLastUpdatedPutResolver
 import eu.kanade.tachiyomi.data.database.resolvers.MangaTitlePutResolver
 import eu.kanade.tachiyomi.data.database.resolvers.SourceIdMangaCountGetResolver
 import eu.kanade.tachiyomi.data.database.tables.ChapterTable
@@ -91,54 +87,23 @@ interface MangaQueries : DbProvider {
 
     fun insertManga(manga: Manga) = db.put().`object`(manga).prepare()
 
-    fun updateChapterFlags(manga: Manga) = db.put()
-        .`object`(manga)
-        .withPutResolver(MangaFlagsPutResolver(MangaTable.COL_CHAPTER_FLAGS, Manga::chapter_flags))
-        .prepare()
-
-    fun updateChapterFlags(manga: List<Manga>) = db.put()
-        .objects(manga)
-        .withPutResolver(MangaFlagsPutResolver(MangaTable.COL_CHAPTER_FLAGS, Manga::chapter_flags, true))
-        .prepare()
-
-    fun updateViewerFlags(manga: Manga) = db.put()
-        .`object`(manga)
-        .withPutResolver(MangaFlagsPutResolver(MangaTable.COL_VIEWER, Manga::viewer_flags))
-        .prepare()
-
-    fun updateViewerFlags(manga: List<Manga>) = db.put()
-        .objects(manga)
-        .withPutResolver(MangaFlagsPutResolver(MangaTable.COL_VIEWER, Manga::viewer_flags, true))
-        .prepare()
-
-    fun updateLastUpdated(manga: Manga) = db.put()
-        .`object`(manga)
-        .withPutResolver(MangaLastUpdatedPutResolver())
-        .prepare()
-
+    // FIXME: Migrate to SQLDelight, on halt: used by StorIO's inTransaction
     fun updateMangaFavorite(manga: Manga) = db.put()
         .`object`(manga)
         .withPutResolver(MangaFavoritePutResolver())
         .prepare()
 
+    // FIXME: Migrate to SQLDelight, on halt: used by StorIO's inTransaction
     fun updateMangaAdded(manga: Manga) = db.put()
         .`object`(manga)
         .withPutResolver(MangaDateAddedPutResolver())
         .prepare()
 
+    // FIXME: Migrate to SQLDelight, on halt: used by StorIO's inTransaction
     fun updateMangaTitle(manga: Manga) = db.put()
         .`object`(manga)
         .withPutResolver(MangaTitlePutResolver())
         .prepare()
-
-    fun updateMangaInfo(manga: Manga) = db.put()
-        .`object`(manga)
-        .withPutResolver(MangaInfoPutResolver())
-        .prepare()
-
-    fun deleteManga(manga: Manga) = db.delete().`object`(manga).prepare()
-
-    fun deleteMangas(mangas: List<Manga>) = db.delete().objects(mangas).prepare()
 
     fun deleteMangasNotInLibraryBySourceIds(sourceIds: List<Long>) = db.delete()
         .byQuery(
@@ -166,14 +131,6 @@ interface MangaQueries : DbProvider {
         )
         .prepare()
 
-    fun deleteMangas() = db.delete()
-        .byQuery(
-            DeleteQuery.builder()
-                .table(MangaTable.TABLE)
-                .build(),
-        )
-        .prepare()
-
     fun getReadNotInLibraryMangas() = db.get()
         .listOfObjects(Manga::class.java)
         .withQuery(
@@ -181,13 +138,5 @@ interface MangaQueries : DbProvider {
                 .query(getReadMangaNotInLibraryQuery())
                 .build(),
         )
-        .prepare()
-
-    fun getTotalChapterManga() = db.get().listOfObjects(Manga::class.java)
-        .withQuery(RawQuery.builder().query(getTotalChapterMangaQuery()).observesTables(MangaTable.TABLE).build()).prepare()
-
-    fun updateMangaFilteredScanlators(manga: Manga) = db.put()
-        .`object`(manga)
-        .withPutResolver(MangaFilteredScanlatorsPutResolver())
         .prepare()
 }