+++ app/src/main/java/yokai/domain/chapter/ChapterRepository.kt
@@ -2,6 +2,7 @@ package yokai.domain.chapter
 
 import eu.kanade.tachiyomi.data.database.models.Chapter
 import kotlinx.coroutines.flow.Flow
+import yokai.domain.chapter.models.ChapterUpdate
 
 interface ChapterRepository {
     suspend fun getChapters(mangaId: Long, filterScanlators: Boolean): List<Chapter>
@@ -9,4 +10,12 @@ interface ChapterRepository {
 
     suspend fun getScanlatorsByChapter(mangaId: Long): List<String>
     fun getScanlatorsByChapterAsFlow(mangaId: Long): Flow<List<String>>
+
+    suspend fun delete(chapter: Chapter): Boolean
+    suspend fun deleteAll(chapters: List<Chapter>): Boolean
+
+    suspend fun update(update: ChapterUpdate): Boolean
+    suspend fun updateAll(updates: List<ChapterUpdate>): Boolean
+
+    suspend fun insert(chapter: Chapter): Long?
 }