+++ app/src/main/java/yokai/domain/chapter/interactor/DeleteChapters.kt
@@ -1,0 +1,11 @@
+package yokai.domain.chapter.interactor
+
+import eu.kanade.tachiyomi.data.database.models.Chapter
+import yokai.domain.chapter.ChapterRepository
+
+class DeleteChapters(
+    private val chapterRepository: ChapterRepository,
+) {
+    suspend fun await(chapter: Chapter) = chapterRepository.delete(chapter)
+    suspend fun awaitAll(chapters: List<Chapter>) = chapterRepository.deleteAll(chapters)
+}