+++ app/src/main/java/yokai/domain/manga/MangaRepository.kt
@@ -6,8 +6,10 @@ import kotlinx.coroutines.flow.Flow
 import yokai.domain.manga.models.MangaUpdate
 
 interface MangaRepository {
-    suspend fun getManga(): List<Manga>
-    fun getMangaAsFlow(): Flow<List<Manga>>
+    suspend fun getMangaList(): List<Manga>
+    suspend fun getMangaByUrlAndSource(url: String, source: Long): Manga?
+    suspend fun getMangaById(id: Long): Manga?
+    fun getMangaListAsFlow(): Flow<List<Manga>>
     suspend fun getLibraryManga(): List<LibraryManga>
     fun getLibraryMangaAsFlow(): Flow<List<LibraryManga>>
     suspend fun update(update: MangaUpdate): Boolean