+++ app/src/main/java/yokai/domain/manga/interactor/GetManga.kt
@@ -1,0 +1,13 @@
+package yokai.domain.manga.interactor
+
+import yokai.domain.manga.MangaRepository
+
+class GetManga (
+    private val mangaRepository: MangaRepository,
+) {
+    suspend fun awaitAll() = mangaRepository.getMangaList()
+    fun subscribeAll() = mangaRepository.getMangaListAsFlow()
+
+    suspend fun awaitByUrlAndSource(url: String, source: Long) = mangaRepository.getMangaByUrlAndSource(url, source)
+    suspend fun awaitById(id: Long) = mangaRepository.getMangaById(id)
+}